site plugin - remove deprecated settings and classes
